Revert "Merge branch 'BugFix-2.1.x' into Creality-Ender-3-Pro-(Mousai)"
This reverts commit 0daf7e547c7a7bd270b93a043bdbce8e2ea4cff7, reversing
changes made to b124e935a4e43701964c0f82abe1150a73229e44.

+++ Marlin/src/lcd/extui/mks_ui/draw_wifi_tips.h
@@ -22,7 +22,7 @@
 #pragma once
 
 #ifdef __cplusplus
-  extern "C" {
+  extern "C" { /* C-declarations for C++ */
 #endif
 
 
@@ -38,7 +38,7 @@ extern TIPS_TYPE wifi_tips_type;
 
 typedef struct {
   unsigned char timer;
-  uint16_t timer_count;
+  unsigned int timer_count;
 } TIPS_DISP;
 extern TIPS_DISP tips_disp;
